Trait engine_traits::KvEngine

pub trait KvEngine: Peekable + SyncMutable + Iterable + WriteBatchExt + DBOptionsExt + CFNamesExt + CFOptionsExt + ImportExt + SstExt + TablePropertiesExt + CompactExt + RangePropertiesExt + MvccPropertiesExt + TtlPropertiesExt + PerfContextExt + MiscExt + Send + Sync + Clone + Debug + 'static {
    type Snapshot: Snapshot;
    fn snapshot(&self) -> Self::Snapshot;

    fn sync(&self) -> Result<()>;

    fn bad_downcast<T: 'static>(&self) -> &T;

    fn flush_metrics(&self, _instance: &str) { ... }

    fn reset_statistics(&self) { ... }
}

A TiKV key-value store

Associated Types

type Snapshot: Snapshot

A consistent read-only snapshot of the database

Loading content...
Required methods

fn snapshot(&self) -> Self::Snapshot

Create a snapshot

fn sync(&self) -> Result<()>

Syncs any writes to disk

fn bad_downcast<T: 'static>(&self) -> &T

Cast to a concrete engine type

This only exists as a temporary hack during refactoring. It cannot be used forever.

Loading content...
Provided methods

fn flush_metrics(&self, _instance: &str)

Flush metrics to prometheus

instance is the label of the metric to flush.

fn reset_statistics(&self)

Reset internal statistics

Loading content...