Trait engine_traits::MiscExt

pub trait MiscExt: Iterable + WriteBatchExt + CFNamesExt {
    fn flush(&self, sync: bool) -> Result<()>;

    fn flush_cf(&self, cf: &str, sync: bool) -> Result<()>;

    fn delete_files_in_range_cf(
        &self, 
        cf: &str, 
        start_key: &[u8], 
        end_key: &[u8], 
        include_end: bool
    ) -> Result<()>;

    fn get_approximate_memtable_stats_cf(
        &self, 
        cf: &str, 
        range: &Range
    ) -> Result<(u64, u64)>;

    fn ingest_maybe_slowdown_writes(&self, cf: &str) -> Result<bool>;

    fn get_engine_used_size(&self) -> Result<u64>;

    fn roughly_cleanup_ranges(
        &self, 
        ranges: &[(Vec<u8>, Vec<u8>)]
    ) -> Result<()>;

    fn path(&self) -> &str;

    fn sync_wal(&self) -> Result<()>;

    fn exists(path: &str) -> bool;

    fn dump_stats(&self) -> Result<String>;

    fn get_latest_sequence_number(&self) -> u64;

    fn get_oldest_snapshot_sequence_number(&self) -> Option<u64>;

    fn is_titan(&self) -> bool { ... }

    fn delete_all_in_range(
        &self, 
        start_key: &[u8], 
        end_key: &[u8], 
        use_delete_range: bool
    ) -> Result<()> { ... }

    fn delete_all_in_range_cf(
        &self, 
        cf: &str, 
        start_key: &[u8], 
        end_key: &[u8], 
        use_delete_range: bool
    ) -> Result<()> { ... }

    fn delete_all_files_in_range(
        &self, 
        start_key: &[u8], 
        end_key: &[u8]
    ) -> Result<()> { ... }
}

Required methods

fn flush(&self, sync: bool) -> Result<()>

fn flush_cf(&self, cf: &str, sync: bool) -> Result<()>

fn delete_files_in_range_cf(
&self,
cf: &str,
start_key: &[u8],
end_key: &[u8],
include_end: bool
) -> Result<()>

fn get_approximate_memtable_stats_cf(
&self,
cf: &str,
range: &Range
) -> Result<(u64, u64)>

Return the approximate number of records and size in the range of memtables of the cf.

fn ingest_maybe_slowdown_writes(&self, cf: &str) -> Result<bool>

fn get_engine_used_size(&self) -> Result<u64>

Gets total used size of rocksdb engine, including:

total size (bytes) of all SST files.

total size (bytes) of active and unflushed immutable memtables.

total size (bytes) of all blob files.

fn roughly_cleanup_ranges(&self, ranges: &[(Vec<u8>, Vec<u8>)]) -> Result<()>

Roughly deletes files in multiple ranges.

Note:

After this operation, some keys in the range might still exist in the database.

After this operation, some keys in the range might be removed from existing snapshot, so you shouldn't expect to be able to read data from the range using existing snapshots any more.

Ref: https://github.com/facebook/rocksdb/wiki/Delete-A-Range-Of-Keys

fn path(&self) -> &str

fn sync_wal(&self) -> Result<()>

fn exists(path: &str) -> bool

Check whether a database exists at a given path

fn dump_stats(&self) -> Result<String>

Dump stats about the database into a string.

For debugging. The format and content is unspecified.

fn get_latest_sequence_number(&self) -> u64

fn get_oldest_snapshot_sequence_number(&self) -> Option<u64>

Loading content...
Provided methods

fn is_titan(&self) -> bool

fn delete_all_in_range(
&self,
start_key: &[u8],
end_key: &[u8],
use_delete_range: bool
) -> Result<()>

fn delete_all_in_range_cf(
&self,
cf: &str,
start_key: &[u8],
end_key: &[u8],
use_delete_range: bool
) -> Result<()>

fn delete_all_files_in_range(
&self,
start_key: &[u8],
end_key: &[u8]
) -> Result<()>

Loading content...