Trait engine_traits::Mutable

pub trait Mutable: Send {
    fn data_size(&self) -> usize;

    fn count(&self) -> usize;

    fn is_empty(&self) -> bool;

    fn should_write_to_engine(&self) -> bool;

    fn clear(&mut self);

    fn set_save_point(&mut self);

    fn pop_save_point(&mut self) -> Result<()>;

    fn rollback_to_save_point(&mut self) -> Result<()>;

    fn put(&mut self, key: &[u8], value: &[u8]) -> Result<()>;

    fn put_cf(&mut self, cf: &str, key: &[u8], value: &[u8]) -> Result<()>;

    fn delete(&mut self, key: &[u8]) -> Result<()>;

    fn delete_cf(&mut self, cf: &str, key: &[u8]) -> Result<()>;

    fn delete_range(&mut self, begin_key: &[u8], end_key: &[u8]) -> Result<()>;

    fn delete_range_cf(
        &mut self, 
        cf: &str, 
        begin_key: &[u8], 
        end_key: &[u8]
    ) -> Result<()>;

    fn put_msg<M: Message>(&mut self, key: &[u8], m: &M) -> Result<()> { ... }

    fn put_msg_cf<M: Message>(
        &mut self, 
        cf: &str, 
        key: &[u8], 
        m: &M
    ) -> Result<()> { ... }
}

A trait implemented by WriteBatch

FIXME: This trait is misnamed, as it is implemented only by WriteBatch and is filled with WriteBatch-specific methods. Merge it into WriteBatch.

Required methods

fn data_size(&self) -> usize

The data size of a write batch

This is necessarily engine-dependent. In RocksDB though it appears to represent the byte length of all write commands in the batch, as serialized in memory, prior to being written to disk.

fn count(&self) -> usize

The number of commands in this batch

fn is_empty(&self) -> bool

Whether any commands have been issued to this batch

fn should_write_to_engine(&self) -> bool

Whether the number of commands exceeds WRITE_BATCH_MAX_KEYS

If so, the write method should be called.

fn clear(&mut self)

Clears the WriteBatch of all commands

It may be reused afterward as an empty batch.

fn set_save_point(&mut self)

Push a save point onto the save point stack

fn pop_save_point(&mut self) -> Result<()>

Pop a save point from the save point stack

This has no effect on the commands already issued to the write batch

fn rollback_to_save_point(&mut self) -> Result<()>

Revert all commands issued since the last save point

Additionally pops the last save point from the save point stack.

fn put(&mut self, key: &[u8], value: &[u8]) -> Result<()>

Write a key/value in the default column family

fn put_cf(&mut self, cf: &str, key: &[u8], value: &[u8]) -> Result<()>

Write a key/value in a given column family

fn delete(&mut self, key: &[u8]) -> Result<()>

Delete a key/value in the default column family

fn delete_cf(&mut self, cf: &str, key: &[u8]) -> Result<()>

Delete a key/value in a given column family

fn delete_range(&mut self, begin_key: &[u8], end_key: &[u8]) -> Result<()>

Delete a range of key/values in the default column family

fn delete_range_cf(
&mut self,
cf: &str,
begin_key: &[u8],
end_key: &[u8]
) -> Result<()>

Delete a range of key/values in a given column family

Loading content...
Provided methods

fn put_msg<M: Message>(&mut self, key: &[u8], m: &M) -> Result<()>

fn put_msg_cf<M: Message>(&mut self, cf: &str, key: &[u8], m: &M) -> Result<()>

Loading content...