Trait engine_traits::SstWriterBuilder

pub trait SstWriterBuilder<E> where
    E: SstExt, {
    fn new() -> Self;

    fn set_db(self, db: &E) -> Self;

    fn set_cf(self, cf: CfName) -> Self;

    fn set_in_memory(self, in_memory: bool) -> Self;

    fn set_compression_type(
        self, 
        compression: Option<SstCompressionType>
    ) -> Self;

    fn set_compression_level(self, level: i32) -> Self;

    fn build(self, path: &str) -> Result<E::SstWriter>;
}

A builder builds a SstWriter.

Required methods

fn new() -> Self

Create a new SstWriterBuilder.

fn set_db(self, db: &E) -> Self

Set DB for the builder. The builder may need some config from the DB.

fn set_cf(self, cf: CfName) -> Self

Set CF for the builder. The builder may need some config from the CF.

fn set_in_memory(self, in_memory: bool) -> Self

Set it to true, the builder builds a in-memory SST builder.

fn set_compression_type(self, compression: Option<SstCompressionType>) -> Self

set other config specified by writer

fn set_compression_level(self, level: i32) -> Self

fn build(self, path: &str) -> Result<E::SstWriter>

Builder a SstWriter.

Loading content...